requirements.txt
torch>=2.0
diffusers>=0.24.0
transformers>=4.38
accelerate>=0.27.2
gradio>=4.16.0
pillow
